Removed all entities except one for development of the API.

+++ src/main/java/uk/co/jacekk/bukkit/entityapi/ai/APINavigation.java
@@ -1,6 +1,6 @@
 package uk.co.jacekk.bukkit.entityapi.ai;
 
-import uk.co.jacekk.bukkit.entityapi.entity.APIEntity;
+import uk.co.jacekk.bukkit.entityapi.entity.api.APIEntity;
 import net.minecraft.server.EntityLiving;
 import net.minecraft.server.MathHelper;
 import net.minecraft.server.Navigation;
@@ -19,7 +19,7 @@ public class APINavigation extends Navigation {
 	
 	@Override
 	public boolean a(double d0, double d1, double d2, float f){
-		PathEntity pathentity = this.a((double) MathHelper.floor(d0), (double) ((int) d1), (double) MathHelper.floor(d2));
+		PathEntity pathentity = this.a(MathHelper.floor(d0), ((int) d1), MathHelper.floor(d2));
 		
 		return this.a(pathentity, f);
 	}
